add a scrool area to widget popup
note: there an error with the layout but I can't figure it out now.
note: I think I mess this up when I decided not to use pointers and heap allocation. This mistake may be at the origin of some of our segfault as Qt may try to delete stuff on the heap (I'm not sure about that, I'll need to investigate)